requirements.txt
fastapi
uvicorn
playwright
pytesseract
pillow
deep-translator
requests
python-multipart